requirements.txt
asgiref==3.2.10
Babel==2.8.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1.1
django-adminlte2==0.4.3
django-appconf==1.0.4
django-bootstrap-modal-forms==2.0.0
django-cities==0.6
django-crispy-forms==1.9.2
django-cruds-adminlte==0.0.16
django-debug-toolbar==3.0
django-formtools==2.2
django-image-cropping==1.4.0
django-model-utils==4.0.0
django-monthfield==0.1.3
django-phonenumber-field==5.0.0
django-rest==0.8.6
django-rest-framework==0.1.0
django-select2==7.4.2
django-weasyprint==1.0.1
django-widget-tweaks==1.4.8
djangoajax==3.2
djangorestframework==3.11.1
easy-thumbnails==2.7
html5lib==1.1
httpie==2.2.0
idna==2.10
phonenumbers==8.12.9
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.1
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
tinycss2==1.0.2
tqdm==4.49.0
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
